requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-bootstrap5==25.1
django-filter==25.1
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
py==1.11.0
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
